Create WIP_pull_request_template.md
When to use it: When you’ve started a new feature, document, or bugfix, that’s not quite ready for others to review, but you want to let your teammates know that you’re working on the feature. This can be used to prevent the duplication of effort, save work that you’ve started, or complement your team’s workflow.